pyproject.toml
[project]
name = "next-gen-analytics"
version = "0.1.0"
description = "RAG demo for home security field notes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "tabulate>=0.10.0",
]
